Mejoras en DietsScreen, UI y lógica de detalle nutricional
- Modifica DietsScreen para que al volver a la pantalla se muestre automáticamente el día actual.
- Ajusta los colores y estilos en ProgressScreen, DietsScreen y AppointmentsScreen para mejorar la coherencia visual.
- Renombra el repositorio de la API a NutritionixRepository e implementa datos de prueba aleatorios.
- Crea FoodDetailViewModel para manejar la lógica de los resultados obtenidos desde la API Nutritionix.
- Modifica la función MealCard en DietsScreen para que navegue a FoodDetailsScreen al hacer clic.

+++ app/src/main/java/com/daniel/mynutricoach/models/FoodInfo.kt
@@ -2,8 +2,8 @@ package com.daniel.mynutricoach.models
 
 data class FoodInfo(
     val name: String = "",
-    val calories: Double = 0.0,
-    val protein: Double = 0.0,
-    val carbs: Double = 0.0,
-    val fat: Double = 0.0,
+    val calories: Float = 0f,
+    val protein: Float = 0f,
+    val carbs: Float = 0f,
+    val fat: Float = 0f
 )
